requirements_termux.txt
aiofiles>=24.1.0
aiohttp>=3.11.8
aiomysql>=0.2.0
aiopath>=0.7.7
aiosqlite>=0.20.0
beautifulsoup4==4.12.3
curl-cffi==0.5.10
ffmpeg-python==0.2.0
jsbeautifier==1.15.1
PyExecJS==1.5.1
pyrofork>=2.3.51
python-dotenv==1.0.1
pyvideothumbnailer==2.1.0
TgCrypto==1.2.5
uvloop==0.19.0
validators>=0.34.0
yt-dlp[default]